added more exceptions
Now the script checks if the number of arguments is correct, if the app you want to start exists at the given location, if the time given can be a number.

+++ watchdog.py
@@ -1,22 +1,15 @@
-"""Realizați un script care să monitorizeze starea unui proces și să îl repornească dacă execuția
-acestuia s-a oprit. Se poate configura intervalul de timp la care să fie verificată starea
-procesului, cât și locația fișierului de log. Se va scrie în log orice modificare a stării procesului(dead / alive). ex: watchdog.py bitcoin_miner.exe 60 watchdog.log"""
+"""
+Author: Formagiu Michael-Gabriel
+This script will be a watchdog(will open it and will make sure it stays up) for an other process.
+"""
 import sys,time,os,subprocess,logging
 
-try:
-	PROCESS_NAME=sys.argv[1]#the process that will be monitorized
-	TIME=sys.argv[2]#in seconds
-	LOGFILE=sys.argv[3]#the logfile that will be created and will write in it
-except:#the default values, if user's input will not be good.
-	PROCESS_NAME="Core Temp.exe"
-	TIME="3"
-	LOGFILE="watchdog.log"
-
 def configLogger():
-	os.remove(LOGFILE)
+	"""In here I will config how log file's format"""
+	if(os.path.exists(LOGFILE)):
+		os.remove(LOGFILE)
 	date_strftime_format = "%d-%b-%y %H:%M:%S"
 	logging.basicConfig(filename=LOGFILE, level=logging.INFO,format='%(asctime)s:%(message)s',datefmt=date_strftime_format)
-configLogger()
 
 def getPID(s):
 	"""This function will get the pid of the app we just opened given a string with the specific format of windows cmd's 'tasklist'."""
@@ -44,6 +37,36 @@ def alive(PID):
 	else:
 		return False
 
+def is_number(s):
+	"""This function checks if a string can be a float number."""
+	try:
+		float(s)
+		return True
+	except:
+		return False
+
+#less arguments than needed
+if (len(sys.argv)<4):
+	LOGFILE="watchdog.log"
+	configLogger()
+	logging.info(' {} arguments are not enough. To run the script you need 4 arguments [script name],[app name],[check cooldown],[logfile name].'.format(len(sys.argv)))
+	exit()
+
+PROCESS_NAME=sys.argv[1]#the process that will be monitorized
+#app not found at given location
+if os.path.exists(PROCESS_NAME)==False:
+	logging.info(' The app with the name: {} could not be found. The script will stop now.'.format(PROCESS_NAME))
+	exit()
+
+TIME=sys.argv[2]#in seconds
+#the argument for time is not in seconds, as needed
+if is_number(TIME)==False:
+	logging.info(' The time given as input: "{}" can not be converted to a number. The script will stop now.'.format(TIME))
+	exit()
+
+LOGFILE=sys.argv[3]#the logfile that will be created and will write in it
+configLogger()
+
 PID=executeApp()#we execute the app for the first time
 logging.info(' {} first started with the PID: {}'.format(PROCESS_NAME,PID))
 while True:
@@ -52,9 +75,3 @@ while True:
 		logging.info(' {} with the PID: {} is not alive anymore. Soon it will start back.'.format(PROCESS_NAME,PID))
 		PID=executeApp()
 		logging.info(' {} came back from the dead. The new PID is: {}'.format(PROCESS_NAME,PID))
-
-#
-
-
-
-
